refactor(shared): move the node:sqlite Effect SQL client into shared
The desktop app needs to read browser cookie databases and should reach for
the same Effect SQL client the server already uses rather than opening
`node:sqlite` by hand. Nothing about the client is server-specific, so it moves
to `packages/shared` and both apps import it from there.

Pure move plus import updates; the client itself is unchanged.
