+++ packages/shared/src/nodeSqliteClient.test.ts
@@ -3,7 +3,7 @@ import * as Effect from "effect/Effect";
 import * as Layer from "effect/Layer";
 import * as SqlClient from "effect/unstable/sql/SqlClient";
 
-import * as SqliteClient from "./NodeSqliteClient.ts";
+import * as SqliteClient from "./nodeSqliteClient.ts";
 
 const layer = it.layer(SqliteClient.layerMemory());
